Update contact.html contact style
The bootstrap style could not respond with the form style but ordinary CSS appear to be fine. May the bootstrap version is also outdated needs to be confirmed. But the problem appeared on the live editor and live browser while on the local machine and editor , the bootstrap functions well.